Google Maps API V3 Viewport Marker Management Snippet
A viewport marker manager works by getting the current bounds of the map
that is visible to the user, then — based on the map's bounds — a query
is sent to the server to get all the markers that lie within the bounds.
As the user pans/zooms the map subsequent requests are sent to the
server to get the new markers. All markers that are no longer in view
are generally removed to improve map performance. --- After Reading many
questions about how to implement this solution for Maps with thousands
of markers., I wrote this code. You can add the code in right after the
map is called and it shouldn't conflict with other listener functions.
Edward Franklin